requirements.txt
psycopg2
sqlalchemy
duckdb
pandas
python-dotenv
faker
dbt
dbt-duckdb
dbt-trino
dbt-postgres